fix response code errors

+++ src/modules/blogger-platform/comments/application/usecases/delete-comment.usecase.ts
@@ -1,10 +1,13 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
-import { NotFoundException } from '@nestjs/common';
+import { ForbiddenException, NotFoundException } from '@nestjs/common';
 import { CommentsRepository } from '../../infrastructure/repositories/comments.repository';
 import { CommentsQueryRepository } from '../../infrastructure/repositories/comments.query.repository';
-
+import { RequestDataEntity } from '../../../../../core/dto/request.data.entity';
 export class DeleteCommentCommand {
-  constructor(public readonly id: string) {}
+  constructor(
+    public readonly id: string,
+    public readonly user: RequestDataEntity,
+  ) {}
 }
 
 @CommandHandler(DeleteCommentCommand)
@@ -17,13 +20,15 @@ export class DeleteCommentUseCase
   ) {}
 
   async execute(command: DeleteCommentCommand): Promise<void> {
-    const { id } = command;
+    const { id, user } = command;
 
     const comment = await this.commentsQueryRepository.findById(id);
     if (!comment) {
       throw new NotFoundException(`Comment with id ${id} not found`);
     }
-
+    if (comment.commentatorInfo.userId !== user.userId) {
+      throw new ForbiddenException();
+    }
     await this.commentsRepository.delete(id);
   }
 }
